[Part 4] Restore the original assistant value from csv in the noop (#10)
Except for the case when the email is invalid either because it's the employee's own email or the assistant records are now gone. They get removed and set to nil.

This PR is setting up so that temp table results will be comparable. Low impact as it's a change to `noop` operations, but needed so that results are similar across both processors.