Add support for non-relation new directory (#14)
In envoy-web, `DirectoryDiff` is used to compare the current directory and the new directory. We will be able to supply DirectoryDiff with a relation version of current directory but not always a relation version of new directory. This is because the initial assumption was that new directory will always be a CSV file (initial thoughts was to use postgres `copy` to slurp the file into a temp table).

While integrating this library, I noticed that we use this library for Google sync, Okta API sync, OneLogin API sync.

This PR keeps all of the previous logic, but adds a step to create a temp table and bulk inserting the array before handing over the relation.