Fix ambiguous projection on current directory relation (#16)
`current_directory` from the outside is a wonderful join of `locations`, `employees` and `assistants`. The projection we pass to `current_directory.select` is unable to disambiguate which table `name` is for. Solution here is to slurp `current_directory` into a temp table and then perform the select on that. This guarantees the column names are unambiguous.